pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "hermes-vault"
version = "0.1.0"
description = "Hermes-native local-first credential broker, scanner, and encrypted vault."
readme = "README.md"
requires-python = ">=3.11"
authors = [
  { name = "Tony Simons" }
]
license = { text = "MIT" }
dependencies = [
  "cryptography>=42.0.0",
  "pydantic>=2.7.0",
  "PyYAML>=6.0.0",
  "rich>=13.7.0",
  "typer>=0.12.0",
  "pathspec>=0.12.1",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0.0",
]

[project.scripts]
hermes-vault = "hermes_vault.cli:app"

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]

[tool.setuptools]
package-dir = { "" = "src" }

[tool.setuptools.packages.find]
where = ["src"]

